Move text processing utilities up a level
